requirements.txt
transformers==5.3.0
torch==2.6.0
langchain==1.2.13
langchain-experimental==0.4.1
langchain-community==0.4.1
langchain-huggingface==1.2.1
PyMuPDF==1.27.2.2
numpy==2.1.3
easyocr==1.7.2
pillow==11.2.1
fastai==2.8.1
uvicorn==0.42.0
